+++ relay-metrics/src/statsd.rs
@@ -48,6 +48,12 @@ pub enum MetricCounters {
     /// This metric is tagged with:
     ///  - `aggregator`: The name of the metrics aggregator (usually `"default"`).
     BucketsDropped,
+    /// Incremented every time the meta aggregator emitted an update that needs to be stored or
+    /// sent upstream.
+    MetaAggregatorUpdate,
+    /// Incremented every time a redis key is updated to store or update metadata.
+    #[cfg(feature = "redis")]
+    MetaRedisUpdate,
 }
 
 impl CounterMetric for MetricCounters {
@@ -56,6 +62,9 @@ impl CounterMetric for MetricCounters {
             Self::MergeHit => "metrics.buckets.merge.hit",
             Self::MergeMiss => "metrics.buckets.merge.miss",
             Self::BucketsDropped => "metrics.buckets.dropped",
+            Self::MetaAggregatorUpdate => "metrics.meta.agg.update",
+            #[cfg(feature = "redis")]
+            Self::MetaRedisUpdate => "metrics.meta.redis.update",
         }
     }
 }
@@ -112,27 +121,11 @@ pub enum MetricHistograms {
     ///    time period (`false`) or after the initial delay has expired (`true`).
     BucketsDelay,
 
-    /// The number of batches emitted per partition by [`crate::aggregator::Aggregator`].
     ///
-    /// This metric is tagged with:
-    ///  - `aggregator`: The name of the metrics aggregator (usually `"default"`).
-    BatchesPerPartition,
-
-    /// The number of buckets in a batch emitted by [`crate::aggregator::Aggregator`].
-    ///
-    /// This corresponds to the number of buckets that will end up in an envelope.
-    ///
-    /// This metric is tagged with:
-    ///  - `aggregator`: The name of the metrics aggregator (usually `"default"`).
-    BucketsPerBatch,
-
     /// Distribution of flush buckets over partition keys.
     ///
     /// The distribution of buckets should be even.
     /// If it is not, this metric should expose it.
-    ///
-    /// This metric is tagged with:
-    ///  - `aggregator`: The name of the metrics aggregator (usually `"default"`).
     PartitionKeys,
 
     /// Distribution of invalid bucket timestamps observed, relative to the time of observation.
@@ -147,8 +140,6 @@ impl HistogramMetric for MetricHistograms {
             Self::BucketsFlushed => "metrics.buckets.flushed",
             Self::BucketsFlushedPerProject => "metrics.buckets.flushed_per_project",
             Self::BucketsDelay => "metrics.buckets.delay",
-            Self::BatchesPerPartition => "metrics.buckets.batches_per_partition",
-            Self::BucketsPerBatch => "metrics.buckets.per_batch",
             Self::PartitionKeys => "metrics.buckets.partition_keys",
             Self::InvalidBucketTimestamp => "metrics.buckets.invalid_timestamp",
         }