+++ relay-server/src/actors/project.rs
@@ -8,7 +8,8 @@ use relay_dynamic_config::{ErrorBoundary, Feature, LimitedProjectConfig, Project
 use relay_filter::matches_any_origin;
 use relay_metrics::aggregator::AggregatorConfig;
 use relay_metrics::{
-    aggregator, Aggregator, Bucket, MergeBuckets, MetricNamespace, MetricResourceIdentifier,
+    aggregator, Aggregator, Bucket, MergeBuckets, MetaAggregator, MetricMeta, MetricNamespace,
+    MetricResourceIdentifier,
 };
 use relay_quotas::{Quota, RateLimits, Scoping};
 use relay_sampling::evaluation::ReservoirCounters;
@@ -19,16 +20,19 @@ use smallvec::SmallVec;
 use tokio::time::Instant;
 use url::Url;
 
-use crate::actors::envelopes::{EnvelopeManager, SendMetrics};
 use crate::actors::outcome::{DiscardReason, Outcome, TrackOutcome};
-use crate::actors::processor::EnvelopeProcessor;
 #[cfg(feature = "processing")]
 use crate::actors::processor::RateLimitBuckets;
+use crate::actors::processor::{EncodeMetricMeta, EnvelopeProcessor};
 use crate::actors::project_cache::{CheckedEnvelope, ProjectCache, RequestUpdate};
 
 use crate::extractors::RequestMeta;
 use crate::statsd::RelayCounters;
-use crate::utils::{EnvelopeLimiter, ManagedEnvelope, MetricsLimiter, RetryBackoff};
+use crate::utils::{
+    EnvelopeLimiter, ExtractionMode, ManagedEnvelope, MetricsLimiter, RetryBackoff,
+};
+
+use super::processor::EncodeMetrics;
 
 /// The expiry status of a project state. Return value of [`ProjectState::check_expiry`].
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
@@ -437,6 +441,8 @@ pub struct Project {
     rate_limits: RateLimits,
     last_no_cache: Instant,
     reservoir_counters: ReservoirCounters,
+    metric_meta_aggregator: MetaAggregator,
+    has_pending_metric_meta: bool,
 }
 
 impl Project {
@@ -452,6 +458,8 @@ impl Project {
             rate_limits: RateLimits::new(),
             last_no_cache: Instant::now(),
             reservoir_counters: Arc::default(),
+            metric_meta_aggregator: MetaAggregator::new(config.metrics_meta_locations_max()),
+            has_pending_metric_meta: false,
             config,
         }
     }
@@ -564,7 +572,8 @@ impl Project {
             _ => false,
         };
 
-        match MetricsLimiter::create(metrics, &state.config.quotas, scoping, usage) {
+        let mode = ExtractionMode::from_usage(usage);
+        match MetricsLimiter::create(metrics, &state.config.quotas, scoping, mode) {
             Ok(mut limiter) => {
                 limiter.enforce_limits(Ok(&self.rate_limits), outcome_aggregator);
                 limiter.into_metrics()
@@ -637,8 +646,9 @@ impl Project {
             Some(ErrorBoundary::Ok(ref c)) => c.usage_metric(),
             _ => false,
         };
+        let extraction_mode = ExtractionMode::from_usage(usage);
 
-        let buckets = match MetricsLimiter::create(buckets, quotas, scoping, usage) {
+        let buckets = match MetricsLimiter::create(buckets, quotas, scoping, extraction_mode) {
             Ok(mut bucket_limiter) => {
                 let cached_rate_limits = self.rate_limits().clone();
                 #[allow(unused_variables)]
@@ -702,6 +712,83 @@ impl Project {
             }
         } else {
             relay_log::debug!("dropping metric buckets, project disabled");
+        }
+    }
+
+    pub fn add_metric_meta(
+        &mut self,
+        meta: MetricMeta,
+        envelope_processor: Addr<EnvelopeProcessor>,
+    ) {
+        let state = self.state_value();
+
+        // Only track metadata if the feature is enabled or we don't know yet whether it is enabled.
+        if !state.map_or(true, |state| state.has_feature(Feature::MetricMeta)) {
+            relay_log::trace!(
+                "metric meta feature flag not enabled for project {}",
+                self.project_key
+            );
+            return;
+        }
+
+        let Some(meta) = self.metric_meta_aggregator.add(self.project_key, meta) else {
+            // Nothing to do. Which means there is also no pending data.
+            relay_log::trace!("metric meta aggregator already has data, nothing to send upstream");
+            return;
+        };
+
+        let scoping = self.scoping();
+        match scoping {
+            Some(scoping) => {
+                // We can only have a scoping if we also have a state, which means at this point feature
+                // flags are already checked.
+                envelope_processor.send(EncodeMetricMeta { scoping, meta })
+            }
+            None => self.has_pending_metric_meta = true,
+        }
+    }
+
+    fn flush_metric_meta(&mut self, envelope_processor: &Addr<EnvelopeProcessor>) {
+        if !self.has_pending_metric_meta {
+            return;
+        }
+        let Some(state) = self.state_value() else {
+            return;
+        };
+        let Some(scoping) = self.scoping() else {
+            return;
+        };
+
+        // All relevant info has been gathered, consider us flushed.
+        self.has_pending_metric_meta = false;
+
+        if !state.has_feature(Feature::MetricMeta) {
+            relay_log::debug!(
+                "clearing metric meta aggregator, because project {} does not have feature flag enabled",
+                self.project_key,
+            );
+            // Project/Org does not have the feature, forget everything.
+            self.metric_meta_aggregator.clear(self.project_key);
+            return;
+        }
+
+        // Flush the entire aggregator containing all code locations for the project.
+        //
+        // There is the rare case when this flushes items which have already been flushed before.
+        // This happens only if we temporarily lose a previously valid and loaded project state
+        // and then reveive an update for the project state.
+        // While this is a possible occurence the effect should be relatively limited,
+        // especially since the final store does de-deuplication.
+        for meta in self
+            .metric_meta_aggregator
+            .get_all_relevant(self.project_key)
+        {
+            relay_log::debug!(
+                "flushing aggregated metric meta for project {}",
+                self.project_key
+            );
+            metric!(counter(RelayCounters::ProjectStateFlushAllMetricMeta) += 1);
+            envelope_processor.send(EncodeMetricMeta { scoping, meta });
         }
     }
 
@@ -913,7 +1000,7 @@ impl Project {
             _ => self.set_state(
                 state.clone(),
                 aggregator,
-                envelope_processor,
+                envelope_processor.clone(),
                 outcome_aggregator,
             ),
         }
@@ -935,6 +1022,14 @@ impl Project {
         relay_log::debug!("project state {} updated", self.project_key);
         channel.inner.send(state);
 
+        self.after_state_updated(&envelope_processor);
+    }
+
+    /// Called after all state validations and after the project state is updated.
+    ///
+    /// See also: [`Self::update_state`].
+    fn after_state_updated(&mut self, envelope_processor: &Addr<EnvelopeProcessor>) {
+        self.flush_metric_meta(envelope_processor);
         // Check if the new sampling config got rid of any reservoir rules we have counters for.
         self.remove_expired_reservoir_rules();
     }
@@ -1015,20 +1110,34 @@ impl Project {
 
     pub fn flush_buckets(
         &mut self,
-        envelope_manager: Addr<EnvelopeManager>,
-        partition_key: Option<u64>,
+        project_cache: Addr<ProjectCache>,
+        envelope_processor: Addr<EnvelopeProcessor>,
         buckets: Vec<Bucket>,
     ) {
+        let Some(project_state) = self.get_cached_state(project_cache, false) else {
+            relay_log::trace!(
+                "there is no project state: dropping {} buckets",
+                buckets.len()
+            );
+            return;
+        };
+
         let Some(scoping) = self.scoping() else {
             relay_log::trace!("there is no scoping: dropping {} buckets", buckets.len());
             return;
         };
 
+        let usage = match project_state.config.transaction_metrics {
+            Some(ErrorBoundary::Ok(ref c)) => c.usage_metric(),
+            _ => false,
+        };
+        let extraction_mode = ExtractionMode::from_usage(usage);
+
         if !buckets.is_empty() {
-            envelope_manager.send(SendMetrics {
+            envelope_processor.send(EncodeMetrics {
                 buckets,
                 scoping,
-                partition_key,
+                extraction_mode,
             });
         }
     }